requirements.txt
# Core chemistry
rdkit>=2023.3.1
chembl-webresource-client>=0.10.8

# Data handling
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.65.0

# Database
sqlalchemy>=2.0.0

# Structure parsing
biopython>=1.81

# Visualization (optional but useful)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
